Refactored DTO's
