+++ DTO/citizenDTO.ts
@@ -1,0 +1,5 @@
+export type CitizenDTO = {
+  id: number;
+  firstName: string;
+  lastName: string;
+};
