+++ DTO/orgoverviewDTO.ts
@@ -1,0 +1,4 @@
+export type OrgOverviewDTO = {
+  id: number;
+  name: string;
+};
