+++ DTO/userDTO.ts
@@ -1,0 +1,6 @@
+export type UserDTO = {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+};
