+++ hooks/useOrganisation.ts
@@ -5,7 +5,8 @@ import {
   deleteMemberRequest,
   fetchOrganisationRequest,
 } from "../apis/organisationAPI";
-import { Citizen, OrgDTO } from "../DTO/organisationDTO";
+import { OrgDTO } from "../DTO/organisationDTO";
+import { CitizenDTO } from "../DTO/citizenDTO";
 
 const useOrganisation = (orgId: number) => {
   const queryClient = useQueryClient();
@@ -16,8 +17,8 @@ const useOrganisation = (orgId: number) => {
     queryKey,
   });
 
-  const createCitizen = useMutation<number, Error, Omit<Citizen, "id">>({
-    mutationFn: (citizen: Omit<Citizen, "id">) =>
+  const createCitizen = useMutation<number, Error, Omit<CitizenDTO, "id">>({
+    mutationFn: (citizen: Omit<CitizenDTO, "id">) =>
       createCitizenRequest(citizen.firstName, citizen.lastName, orgId),
     onMutate: async (newCitizen) => {
       await queryClient.cancelQueries({ queryKey });
