+++ hooks/useProfile.ts
@@ -1,19 +1,7 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useAuthentication } from "../providers/AuthenticationProvider";
 import { changePasswordRequest, fetchProfileRequest, updateProfileRequest } from "../apis/profileAPI";
-
-export type ProfileDTO = {
-  email: string;
-  firstName: string;
-  lastName: string;
-};
-
-export type UpdateProfileDTO = Omit<ProfileDTO, "email">;
-
-export type ChangePasswordDTO = {
-  oldPassword: string;
-  newPassword: string;
-};
+import { ProfileDTO, UpdateProfileDTO, ChangePasswordDTO } from "../DTO/ProfileDTO";
 
 export default function useProfile() {
   const { userId } = useAuthentication();
